Remove offersCredit from BTPayPalRequest and move userAction to BTPayPalCheckoutRequest
* Convert PayPal Credit demo into a Pay Later demo
