--- Demo/Application/Features/PayPal - Credit/BraintreeDemoPayPalCreditPaymentViewController.h
@@ -1,5 +1,0 @@
-#import "BraintreeDemoPaymentButtonBaseViewController.h"
-
-@interface BraintreeDemoPayPalCreditPaymentViewController : BraintreeDemoPaymentButtonBaseViewController
-
-@end
